Fixed multiple classes defined in a single file.
