+++ src/ClientException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PayPay\OpenPaymentAPI;
+
+use Exception;
+
+class ClientException extends Exception
+{
+}
